+++ packages/expo-location/src/LocationSubscribers.ts
@@ -83,6 +83,7 @@ export const HeadingSubscriber = new Subscriber<LocationHeadingCallback>(
 
 /**
  * Necessary for some unit tests.
+ * @returns The current subscription id
  */
 export function _getCurrentWatchId(): number {
   return nextWatchId;